fix the issue that xcode cannot view the device process list, also "frida-ps -U"
